Accept Go toolchain parameter
runtime.GOROOT() is the GOROOT for the toolchain used to build
go-patch-overlay. This will probably be the same as the toolchain used to build
the program where the overlay will be used. But it's a little more flexible to
have this tool accept a path to a toolchain. For example, it makes the tool
easier to test. The user might also get a compiled binary of this tool from
elsewhere (package manager?) or might try to build their program with a
toolchain other than their "default" one (e.g. golang.org/dl/go1.x.y).