Add unit tests
The unit tests consist of:

- A toolchain "path"
- Patches
- A Go program
- Expected output from the program after building with the patches

The tests create an overlay using the patches for the given toolchain, and then
build the program with that overlay, run it, and check its output. The test use
golang.org/x/txtar to present this collection of data. I chose this because I
think it lets us have a nice, concise representation of a test case. My hope is
that this format will make it very easy to add new tests. The tests extract
these "archives" into a temporary directory and run the tool (basically via the
Overlay struct). We could even try building the go-patch-overlay binary and
running that if we want even more end-to-end tests.